+++ context.go
@@ -69,6 +69,7 @@ const (
 	keyRefillIndexCaches        ContextKey = "arangodb-driver-refill-index-caches"
 	keyAsyncRequest             ContextKey = "arangodb-async-request"
 	keyAsyncID                  ContextKey = "arangodb-async-id"
+	keySkipIfExistCheck         ContextKey = "arangodb-skip-if-exist-check"
 )
 
 type OverwriteMode string
@@ -305,6 +306,12 @@ func WithAsync(parent context.Context) context.Context {
 // WithAsyncID is used to check an async operation result - requires Connection with Async wrapper!
 func WithAsyncID(parent context.Context, asyncID string) context.Context {
 	return context.WithValue(contextOrBackground(parent), keyAsyncID, asyncID)
+}
+
+// WithSkipIfExistCheck is used to disable validation for resource existence
+// e.g.: ClientDatabases.Database will do not call the additional check to ArangoDB for ensuring that DB exist
+func WithSkipIfExistCheck(parent context.Context, value bool) context.Context {
+	return context.WithValue(contextOrBackground(parent), keySkipIfExistCheck, value)
 }
 
 type contextSettings struct {