+++ v2/tests/database_collection_operations_test.go
@@ -26,13 +26,27 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"golang.org/x/text/unicode/norm"
-
 	"github.com/arangodb/go-driver/v2/arangodb"
 	"github.com/arangodb/go-driver/v2/arangodb/shared"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
+
+func TestGetCollection(t *testing.T) {
+	Wrap(t, func(t *testing.T, c arangodb.Client) {
+		WithDatabase(t, c, nil, func(db arangodb.Database) {
+			withContextT(t, defaultTestTimeout, func(ctx context.Context, _ testing.TB) {
+				// The collection should not be found
+				_, err := db.GetCollection(ctx, "wrong-name", nil)
+				require.NotNil(t, err)
+
+				// IsExist validation should be skipped
+				_, err = db.GetCollection(ctx, "wrong-name", &arangodb.GetCollectionOptions{SkipExistCheck: true})
+				require.Nil(t, err)
+			})
+		})
+	})
+}
 
 // Test_CollectionShards creates a collection and gets the shards' information.
 func Test_CollectionShards(t *testing.T) {
@@ -462,68 +476,6 @@ func Test_DatabaseCollectionOperations(t *testing.T) {
 			})
 		})
 	})
-}
-
-func TestDatabaseNameUnicode(t *testing.T) {
-	Wrap(t, func(t *testing.T, c arangodb.Client) {
-		withContextT(t, defaultTestTimeout, func(ctx context.Context, _ testing.TB) {
-			skipBelowVersion(c, ctx, "3.9.0", t)
-			databaseExtendedNamesRequired(t, c, ctx)
-
-			random := GenerateUUID("test-db-unicode")
-			dbName := "\u006E\u0303\u00f1" + random
-			_, err := c.CreateDatabase(ctx, dbName, nil)
-			require.EqualError(t, err, "database name is not properly UTF-8 NFC-normalized")
-
-			normalized := norm.NFC.String(dbName)
-			_, err = c.CreateDatabase(ctx, normalized, nil)
-			require.NoError(t, err)
-
-			// The database should not be found by the not normalized name.
-			_, err = c.Database(ctx, dbName)
-			require.NotNil(t, err)
-
-			// The database should be found by the normalized name.
-			exist, err := c.DatabaseExists(ctx, normalized)
-			require.NoError(t, err)
-			require.True(t, exist)
-
-			var found bool
-			databases, err := c.Databases(ctx)
-			require.NoError(t, err)
-			for _, database := range databases {
-				if database.Name() == normalized {
-					found = true
-					break
-				}
-			}
-			require.Truef(t, found, "the database %s should have been found", normalized)
-
-			// The database should return handler to the database by the normalized name.
-			db, err := c.Database(ctx, normalized)
-			require.NoError(t, err)
-			require.NoErrorf(t, db.Remove(ctx), "failed to remove testing database")
-		})
-	})
-}
-
-// databaseExtendedNamesRequired skips test if the version is < 3.9.0 or the ArangoDB has not been launched
-// with the option --database.extended-names-databases=true.
-func databaseExtendedNamesRequired(t *testing.T, c arangodb.Client, ctx context.Context) {
-	// If the database can be created with the below name then it means that it excepts unicode names.
-	dbName := "\u006E\u0303\u00f1" + GenerateUUID("test-db")
-	normalized := norm.NFC.String(dbName)
-	db, err := c.CreateDatabase(ctx, normalized, nil)
-	if err == nil {
-		require.NoErrorf(t, db.Remove(ctx), "failed to remove testing database")
-	}
-
-	if shared.IsArangoErrorWithErrorNum(err, shared.ErrArangoDatabaseNameInvalid, shared.ErrArangoIllegalName) {
-		t.Skipf("ArangoDB is not launched with the option --database.extended-names-databases=true")
-	}
-
-	// Some other error which has not been expected.
-	require.NoError(t, err)
 }
 
 func Test_DatabaseCollectionTruncate(t *testing.T) {