+++ v2/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [master](https://github.com/arangodb/go-driver/tree/master) (N/A)
 - Switch to Go 1.21.5
 - Disable AF mode in tests (not supported since 3.12)
+- Allow skipping validation for Database and Collection existence
 
 ## [2.0.3](https://github.com/arangodb/go-driver/tree/v2.0.3) (2023-10-31)
 - Add optional status code checks. Consistent return of response